调整修改

+++ diyi-service/diyi-order/src/main/java/com/lgyun/system/order/controller/admin/EnterpriseAdminController.java
@@ -2,12 +2,15 @@ package com.lgyun.system.order.controller.admin;
 
 import com.lgyun.common.api.R;
 import com.lgyun.common.enumeration.ObjectType;
+import com.lgyun.common.secure.BladeUser;
 import com.lgyun.common.tool.Func;
 import com.lgyun.core.mp.support.Condition;
 import com.lgyun.core.mp.support.Query;
 import com.lgyun.system.order.dto.AddressDTO;
 import com.lgyun.system.order.service.IAddressService;
 import com.lgyun.system.order.service.IPayEnterpriseService;
+import com.lgyun.system.user.entity.AdminEntity;
+import com.lgyun.system.user.feign.IUserClient;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
@@ -32,30 +35,56 @@ import javax.validation.constraints.NotNull;
 @Api(value = "平台端---商户管理模块相关接口", tags = "平台端---商户管理模块相关接口")
 public class EnterpriseAdminController {
 
+    private IUserClient userClient;
     private IAddressService addressService;
     private IPayEnterpriseService payEnterpriseService;
 
     @GetMapping("/query-address-list")
     @ApiOperation(value = "查询商户所有收货地址信息", notes = "查询商户所有收货地址信息")
-    public R queryAddressList(@ApiParam(value = "商户ID") @NotNull(message = "请输入商户编号") @RequestParam(required = false) Long enterpriseId, Query query) {
+    public R queryAddressList(@ApiParam(value = "商户ID") @NotNull(message = "请输入商户编号") @RequestParam(required = false) Long enterpriseId, Query query, BladeUser bladeUser) {
+        //查询当前管理员
+        R<AdminEntity> result = userClient.currentAdmin(bladeUser);
+        if (!(result.isSuccess())) {
+            return result;
+        }
+
         return addressService.queryAddressList(ObjectType.ENTERPRISEPEOPLE, enterpriseId, Condition.getPage(query.setDescs("create_time")));
     }
 
     @PostMapping("/add-or-update-address")
     @ApiOperation(value = "添加/编辑收货地址", notes = "添加/编辑收货地址")
-    public R addOrUpdateAddress(@ApiParam(value = "商户编号") @NotNull(message = "请选择商户") @RequestParam(required = false) Long enterpriseId, @Valid @RequestBody AddressDTO addressDto) {
+    public R addOrUpdateAddress(@ApiParam(value = "商户编号") @NotNull(message = "请选择商户") @RequestParam(required = false) Long enterpriseId,
+                                @Valid @RequestBody AddressDTO addressDto, BladeUser bladeUser) {
+        //查询当前管理员
+        R<AdminEntity> result = userClient.currentAdmin(bladeUser);
+        if (!(result.isSuccess())) {
+            return result;
+        }
+
         return addressService.addOrUpdateAddress(addressDto, enterpriseId, ObjectType.ENTERPRISEPEOPLE);
     }
 
     @PostMapping("/remove-address")
     @ApiOperation(value = "删除收货地址", notes = "删除收货地址")
-    public R removeAddress(@ApiParam(value = "收货地址ID集合", required = true) @NotBlank(message = "请选择要删除的收货地址") @RequestParam(required = false) String ids) {
+    public R removeAddress(@ApiParam(value = "收货地址ID集合", required = true) @NotBlank(message = "请选择要删除的收货地址") @RequestParam(required = false) String ids, BladeUser bladeUser) {
+        //查询当前管理员
+        R<AdminEntity> result = userClient.currentAdmin(bladeUser);
+        if (!(result.isSuccess())) {
+            return result;
+        }
+
         return R.status(addressService.removeByIds(Func.toLongList(ids)));
     }
 
     @GetMapping("/enterprise-transaction")
     @ApiOperation(value = "查询商户交易数据", notes = "查询商户交易数据")
-    public R transactionByEnterprise(@ApiParam(value = "商户ID") @NotNull(message = "请输入商户编号") @RequestParam(required = false) Long enterpriseId) {
+    public R transactionByEnterprise(@ApiParam(value = "商户ID", required = true) @NotNull(message = "请输入商户编号") @RequestParam(required = false) Long enterpriseId, BladeUser bladeUser) {
+        //查询当前管理员
+        R<AdminEntity> result = userClient.currentAdmin(bladeUser);
+        if (!(result.isSuccess())) {
+            return result;
+        }
+
         return payEnterpriseService.transactionByEnterprise(enterpriseId);
     }
